+++ menger-app/src/test/scala/menger/cli/CLIObjectOptionSuite.scala
@@ -6,30 +6,22 @@ import org.scalatest.matchers.should.Matchers
 
 class CLIObjectOptionSuite extends AnyFlatSpec with Matchers:
 
-  "--objects" should "accept 'type=sphere' with --optix" in:
-    val options = SafeMengerCLIOptions(Seq("--optix", "--objects", "type=sphere"))
+  "--objects" should "accept 'type=sphere'" in:
+    val options = SafeMengerCLIOptions(Seq("--objects", "type=sphere"))
     options.objects.toOption.flatMap(_.headOption.map(_.objectType)) shouldBe Some("sphere")
 
-  it should "accept 'type=cube' with --optix" in:
-    val options = SafeMengerCLIOptions(Seq("--optix", "--objects", "type=cube"))
+  it should "accept 'type=cube'" in:
+    val options = SafeMengerCLIOptions(Seq("--objects", "type=cube"))
     options.objects.toOption.flatMap(_.headOption.map(_.objectType)) shouldBe Some("cube")
-
-  it should "reject without --optix" in:
-    an[ScallopException] should be thrownBy:
-      SafeMengerCLIOptions(Seq("--objects", "type=sphere"))
 
   it should "reject unknown object type" in:
     an[ScallopException] should be thrownBy:
-      SafeMengerCLIOptions(Seq("--optix", "--objects", "type=pyramid"))
+      SafeMengerCLIOptions(Seq("--objects", "type=pyramid"))
 
-  "--optix" should "require --objects option" in:
-    an[ScallopException] should be thrownBy:
-      SafeMengerCLIOptions(Seq("--optix"))
-
-  it should "work with --objects type=sphere" in:
+  "--objects type=sphere" should "work standalone" in:
     noException should be thrownBy:
-      SafeMengerCLIOptions(Seq("--optix", "--objects", "type=sphere"))
+      SafeMengerCLIOptions(Seq("--objects", "type=sphere"))
 
-  it should "work with --objects type=cube" in:
+  it should "work with type=cube" in:
     noException should be thrownBy:
-      SafeMengerCLIOptions(Seq("--optix", "--objects", "type=cube"))
+      SafeMengerCLIOptions(Seq("--objects", "type=cube"))